backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.26.0
openai==1.10.0
anthropic==0.18.1


pydantic==2.6.0
pydantic-settings==2.1.0
tiktoken==0.5.2
beautifulsoup4==4.12.3
requests==2.31.0
duckduckgo-search==4.4.1
termcolor==2.4.0

python-docx==1.1.0
